requirements.txt
google-adk
google-genai
python-dotenv
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
yfinance
pandas
pydantic